refactor(obi-390): refactored some functions to prevent code duplication, fixed lastMessage
